Moves EditableText, EntityCard, Flag, FlashMessage
