+++ amundsen_application/static/js/components/EntityCard/EntityCardSection/styles.scss
