+++ amundsen_application/static/js/components/EntityCard/styles.scss
