+++ amundsen_application/static/js/components/Flag/styles.scss
