+++ amundsen_application/static/js/components/FlashMessage/flashMessage.story.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 
 import { ImageIconType } from 'interfaces/Enums';
-import StorySection from '../StorySection';
+import StorySection from '../common/StorySection';
 import FlashMessage from '.';
 
 const stories = storiesOf('Components/Flash Message', module);
