+++ amundsen_application/static/js/pages/TableDetailPage/RequestMetadataForm/index.tsx
@@ -5,8 +5,6 @@ import * as React from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
-import './styles.scss';
-
 import {
   NotificationType,
   RequestMetadataType,
@@ -15,7 +13,7 @@ import {
   TableMetadata,
 } from 'interfaces';
 
-import FlashMessage from 'components/common/FlashMessage';
+import FlashMessage from 'components/FlashMessage';
 import { ImageIconType } from 'interfaces/Enums';
 
 import { GlobalState } from 'ducks/rootReducer';
@@ -29,6 +27,8 @@ import {
   submitNotification,
 } from 'ducks/notification/reducer';
 import * as Constants from './constants';
+
+import './styles.scss';
 
 interface StateFromProps {
   columnName?: string;
