+++ amundsen_application/static/js/pages/TableDetailPage/index.tsx
@@ -32,7 +32,7 @@ import BookmarkIcon from 'components/Bookmark/BookmarkIcon';
 import Breadcrumb from 'components/Breadcrumb';
 import TabsComponent, { TabInfo } from 'components/common/TabsComponent';
 import TagInput from 'components/common/Tags/TagInput';
-import EditableText from 'components/common/EditableText';
+import EditableText from 'components/EditableText';
 import LoadingSpinner from 'components/common/LoadingSpinner';
 import EditableSection from 'components/EditableSection';
 import ColumnList from 'features/ColumnList';
